split deprecated constructors into deprecated.go; expand README docs
polyscript.go was 478 lines; the bulk of that was 12 deprecated FromXxx*
constructors that crowded the new API. Move them into deprecated.go in
the same package so the canonical API in polyscript.go is what readers
land on first.

README updates:

- Trim the Overview's "tiers of public APIs" wording; the canonical API
  is now polyscript.New[E].
- Add an "API at a glance" section with three small tables: Engines,
  Sources, and Options. Each entry links the public name to its purpose
  and notes the compile-time engine binding for WithEntryPoint.
- Promote the type-argument note next to the Options table where users
  will look for it.
- Add a "Loading Scripts" section with worked examples for FromFile, an
  HTTP loader via FromLoader (timeout + bearer auth), and a logger
  handler example.

No source-level behavior change; just file-organization and docs.
